- Removing _trx_txn member variable from trx_packet - Adding bdb_trx_packet class that derives from trx_packet and contains the   BDB-specific _trx_txn member varialbe - Changing all the BDB-related staged to use the new bdb_trx_packet class

+++ src/workload/tpcc/dbgen/tpcc_rnd.cpp
@@ -81,10 +81,10 @@ int random_1_3000( void ) {
 
 // fn: initialize_random
 void initialize_random(void) {
-  int t = current_time();                                     
+    int t = (int)current_time();                                     
 
-  srand(t);                                                   
-  srandom(t);                                                 
+    srand(t);                                                   
+    srandom(t);                                                 
 }
